gcc fixes

+++ src/fpoint.cpp
@@ -65,9 +65,9 @@ float angle_radians(const fpoint &p)
     if (p.y > 0) {
       return theta;
     }
-    return (RAD_360 + theta);
+    return ((float) RAD_360 + theta);
   }
-  return (RAD_180 - theta);
+  return ((float) RAD_180 - theta);
 }
 
 fpoint rotate_radians(float angle, const fpoint &p, const fpoint &O)
